+++ web/static/react/assets/BackupPage-Bm6Uu_n2.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-D87v-pXH.js";import{t}from"./ImpostazioniPage-CDgmxFsA.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as BackupPage};