+++ web/static/react/assets/PracticeProfilePage-CPDFB19Q.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-D87v-pXH.js";import{t}from"./LegalSkillsProfilePage-DfrJTI7B.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as PracticeProfilePage};