+++ web/static/react/assets/PracticeProfilePage-DObvI0QP.js
@@ -1,0 +1,1 @@
+import{t as e}from"./vendor-react-D87v-pXH.js";import{t}from"./LegalSkillsProfilePage-D3VmGOA3.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as PracticeProfilePage};