+++ web/static/react/assets/SkillRunDetailPage-pR1C0iLI.js
@@ -1,1 +1,0 @@
-import{t as e}from"./vendor-react-D87v-pXH.js";import{t}from"./LegalSkillsReviewPage-B9GqN-iS.js";var n=e();function r(){return(0,n.jsx)(t,{})}export{r as SkillRunDetailPage};